Fix: add security

+++ src/users/users.service.ts
@@ -1,9 +1,10 @@
 import handleErrorConstraintUnique from "../utils/handle-error-unique.util";
-import { Injectable, NotFoundException } from "@nestjs/common";
+import { Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
 import { PrismaService } from "src/prisma/prisma.service";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import * as bcrypt from "bcryptjs";
+import { Users } from "./entities/users.entities";
 
 @Injectable()
 export class UsersService {
@@ -11,7 +12,6 @@ export class UsersService {
 		id: true,
 		name: true,
 		email: true,
-		cpf: true,
 		updatedAt: true,
 		createdAt: true,
 	};
@@ -34,10 +34,7 @@ export class UsersService {
 
 	async findAll() {
 		const res = await this.prisma.users.findMany({
-			select: {
-				...this.userSelect,
-				profile: true,
-			},
+			select: this.userSelect,
 		});
 
 		return res;
@@ -48,6 +45,8 @@ export class UsersService {
 			where: { id },
 			select: {
 				...this.userSelect,
+				isAdmin: true,
+				cpf: true,
 				profile: true,
 			},
 		});
@@ -63,23 +62,38 @@ export class UsersService {
 		return await this.verifyIdAndReturnUser(id);
 	}
 
-	async update(id: string, dto: UpdateUserDto) {
-		await this.verifyIdAndReturnUser(id);
+	async update(id: string, dto: UpdateUserDto, user: Users) {
+		const thisUser = await this.verifyIdAndReturnUser(id);
 
 		if (dto.password) {
 			const hashedPassword = await bcrypt.hash(dto.password, 8);
 			dto.password = hashedPassword;
 		}
-
-		return await this.prisma.users.update({ where: { id }, data: dto, select: this.userSelect }).catch(handleErrorConstraintUnique);
+		if (user.isAdmin) {
+			return await this.prisma.users.update({ where: { id }, data: dto, select: this.userSelect }).catch(handleErrorConstraintUnique);
+		}
+		if (thisUser.id === user.id) {
+			return await this.prisma.users.update({ where: { id }, data: dto, select: this.userSelect }).catch(handleErrorConstraintUnique);
+		} else {
+			throw new UnauthorizedException("not authorized");
+		}
 	}
 
-	async remove(id: string) {
-		await this.verifyIdAndReturnUser(id);
-
-		return await this.prisma.users.delete({
-			where: { id },
-			select: this.userSelect,
-		});
+	async remove(id: string, user: Users) {
+		const thisUser = await this.verifyIdAndReturnUser(id);
+		if (user.isAdmin) {
+			return await this.prisma.users.delete({
+				where: { id },
+				select: this.userSelect,
+			});
+		}
+		if (thisUser.id === user.id) {
+			return await this.prisma.users.delete({
+				where: { id },
+				select: this.userSelect,
+			});
+		} else {
+			throw new UnauthorizedException("not authorized");
+		}
 	}
 }
